Separate sound from foundational wrapper
