+++ client/engine/backend/README.md
@@ -1,0 +1,4 @@
+# Backend
+
+The backend is an abstraction over the underlying libraries that support sound, networking, etc...
+Currently all of them are based on pygame.
