+++ client/engine/backend/foundational_wrapper.py
@@ -49,26 +49,6 @@ class FoundationalWrapper:
         return pygame.display.set_mode((width, height))
 
     @staticmethod
-    def play_sound(path: str) -> None:
-        pygame.mixer.Sound.play(pygame.mixer.Sound(path))
-        return None
-
-    @staticmethod
-    def load_music(path: str) -> None:
-        pygame.mixer.music.load(path)
-        return None
-
-    @staticmethod
-    def play_music() -> None:
-        pygame.mixer.music.play(-1)
-        return None
-
-    @staticmethod
-    def stop_music() -> None:
-        pygame.mixer.music.stop()
-        return None
-
-    @staticmethod
     def get_clock_ticks() -> int:
         return int(pygame.time.get_ticks())
