+++ client/engine/features/sound/event_handler.py
@@ -1,7 +1,7 @@
 import logging
 from typing import TYPE_CHECKING, Any, Dict, Type
 
-from client.engine.external.foundational_wrapper import FoundationalWrapper
+from client.engine.backend.sound import SoundBackend
 from client.engine.primitives.event_handler import EventHandler
 
 from .events import PlayMusicEvent, PlaySoundEvent, StopMusicEvent
@@ -16,20 +16,20 @@ class PlaySoundEventHandler(EventHandler[PlaySoundEvent]):
     def handle(self, event: "PlaySoundEvent") -> None:
         # TODO: Event.sound is an id, not a path
         # but the sound module expects a path.
-        FoundationalWrapper.play_sound(event.sound)
+        SoundBackend.play_sound(event.sound)
 
 
 class PlayMusicEventHandler(EventHandler[PlayMusicEvent]):
     def handle(self, event: "PlayMusicEvent") -> None:
         # TODO: Event.music is an id, not a path
         # but the music module expects a path.
-        FoundationalWrapper.load_music(event.music)
-        FoundationalWrapper.play_music()
+        SoundBackend.load_music(event.music)
+        SoundBackend.play_music()
 
 
 class StopMusicEventHandler(EventHandler[StopMusicEvent]):
     def handle(self, event: "StopMusicEvent") -> None:
-        FoundationalWrapper.stop_music()
+        SoundBackend.stop_music()
 
 
 handlers_map: Dict[Type["Event"], Any] = {
