+++ client/engine/features/user_input/keyboard.py
@@ -1,6 +1,6 @@
 from typing import Any, List
 
-from client.engine.external.foundational_wrapper import FoundationalWrapper
+from client.engine.backend.foundational_wrapper import FoundationalWrapper
 
 
 class KeyboardInput:
