+++ client/engine/features/user_input/mouse.py
@@ -1,6 +1,6 @@
 from typing import Any, List, Optional
 
-from client.engine.external.foundational_wrapper import FoundationalWrapper
+from client.engine.backend.foundational_wrapper import FoundationalWrapper
 
 
 class MouseInput:
