+++ client/feature_tests/network.py
@@ -1,6 +1,5 @@
 from client.engine.features.network.commands import SendNetworkRequest
 from client.engine.features.network.worker import NetworkWorker
-from client.engine.general_state.options import Options
 from client.engine.general_state.queue import QueueManager
 
 
@@ -17,7 +16,6 @@ def on_error(event):
 
 if __name__ == "__main__":
     QueueManager().initialize()
-    Options().initialize()
     network_thread = NetworkWorker(
         name="Network",
         queue=QueueManager().get("network"),
