+++ run_network_feature_test
@@ -1,2 +1,2 @@
 #!/bin/bash
-poetry run python -m client.engine.feature_tests.network
+poetry run python -m client.feature_tests.network
