+++ run_sound_feature_test
@@ -1,2 +1,2 @@
 #!/bin/bash
-poetry run python -m client.engine.feature_tests.sound
+poetry run python -m client.feature_tests.sound
