+++ run_user_input_feature_test
@@ -1,2 +1,2 @@
 #!/bin/bash
-poetry run python -m client.engine.feature_tests.user_input
+poetry run python -m client.feature_tests.user_input
